requirements.txt
Flask==2.3.0
Flask-SQLAlchemy==3.1.1
Werkzeug==2.3.0
pymysql==1.1.0
python-dotenv==1.0.1